LayoutTests/imported/w3c/web-platform-tests/svg/path/error-handling/bounding-expected.txt

CONSOLE MESSAGE: Error: Problem parsing d="none"
CONSOLE MESSAGE: Error: Problem parsing d="# invalid"
CONSOLE MESSAGE: Error: Problem parsing d="m 0 0 l 3 -4 z # ignored suffix v 123"
CONSOLE MESSAGE: Error: Problem parsing d="m 0 0 l -9 11 -123 z # ignore last l parameter"

PASS bounding
